requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
pydantic>=2.0,<3.0
aiofiles>=23.0.0
Pillow>=10.0.0
pillow-heif>=0.13.0
piexif>=1.1.3
